Add first pass at tests for fold pparams
Uses the `crawler` example from pallas to grab a bunch of mainnet update proposals.
Tests that folding to each of these epochs works and produces the correct protocol version.

The assertion here is just on the protocol version, with a pending todo to test this more
thoroughly once we can serialize / deserialize the interim protocol parameter files.